update to script

+++ deploy/scripts/build_application.sh
@@ -2,13 +2,28 @@
 
 echo "Creating the IdeaJet Application"
 
+while getopts ":e:" opt; do
+  case $opt in
+    e)
+      environment=$OPTARG
+      ;;
+    *)
+      echo 'Error in command line parsing' >&2
+      exit 1
+  esac
+done
+
+if [ -z "$environment" ]; then
+        echo 'Missing -e' >&2
+        exit 1
+fi
 
 rm -rf /applications/ideajet/
 mkdir /applications/ideajet/
 
 echo "BUILDING"
 
-package_name=ideajet
+package_name="${environment}_ideajet"
 jenkins_proj_path="/var/lib/jenkins/workspace/$package_name"
 JENKINS_VENV_DIR=$jenkins_proj_path/venv 
 
@@ -28,7 +43,7 @@ python -m venv /applications/ideajet/venv
 . "/applications/ideajet/venv/bin/activate"
 pip install --upgrade pip
 pip install wheel
-pip install /var/lib/jenkins/workspace/int_ideajet/dist/ideajet-0.1.0-py3-none-any.whl
+pip install "/var/lib/jenkins/workspace/${package_name}/dist/ideajet-0.1.0-py3-none-any.whl"
 echo "Application packages installed into Venv"
 
 echo "Gzipping Application"
